Move mock implementations into specific protocols.
